src/cloud-compute/requirements.txt
flask
openai-whisper
gunicorn
numpy